extracted deprecated projectOpened listener to ProjectActivity
