+++ src/main/liveplugin/implementation/LivePluginProjectPostStartupActivity.kt
@@ -1,0 +1,29 @@
+package liveplugin.implementation
+
+import com.intellij.ide.impl.isTrusted
+import com.intellij.notification.NotificationType
+import com.intellij.openapi.actionSystem.ActionManager
+import com.intellij.openapi.actionSystem.AnActionEvent
+import com.intellij.openapi.actionSystem.CommonDataKeys
+import com.intellij.openapi.actionSystem.Presentation
+import com.intellij.openapi.project.Project
+import com.intellij.openapi.startup.ProjectActivity
+import liveplugin.implementation.common.MapDataContext
+import liveplugin.implementation.common.livePluginNotificationGroup
+import liveplugin.implementation.pluginrunner.PluginRunner.Companion.runPlugins
+
+class LivePluginProjectPostStartupActivity : ProjectActivity {
+    override suspend fun execute(project: Project) {
+        if (!Settings.instance.runProjectSpecificPlugins) return
+        @Suppress("UnstableApiUsage")
+        if (!project.isTrusted()) {
+            val message = "Skipped execution of project specific plugins because the project is not trusted."
+            livePluginNotificationGroup.createNotification(title = "Live plugin", message, NotificationType.INFORMATION).notify(project)
+            return
+        }
+
+        val dataContext = MapDataContext(mapOf(CommonDataKeys.PROJECT.name to project))
+        val dummyEvent = AnActionEvent(null, dataContext, "", Presentation(), ActionManager.getInstance(), 0)
+        runPlugins(readLivePlugins(project), dummyEvent)
+    }
+}
