+++ src/main/liveplugin/implementation/ProjectPluginsReader.kt
@@ -1,0 +1,9 @@
+package liveplugin.implementation
+
+import com.intellij.openapi.project.Project
+import liveplugin.implementation.common.toFilePath
+
+fun readLivePlugins(project: Project): List<LivePlugin> {
+    val projectPath = project.basePath?.toFilePath() ?: return emptyList()
+    return (projectPath + LivePluginPaths.livePluginsProjectDirName).listFiles().toLivePlugins()
+}
